FAR patch 00649.MoreNWZ Дата              : 12.05.2001 Сделал            : Dmitry Jemerov Описание          : Some more NWZ Измененные файлы  : chgmmode.cpp cmdline.cpp config.cpp copy.cpp ctrlobj.cpp                     ctrlobj.hpp delete.cpp dialog.cpp edit.cpp editor.cpp                     far.dsp ffolders.cpp fileedit.cpp filelist.cpp                     filepanels.cpp filetype.cpp fileview.cpp filter.cpp                     findfile.cpp flmodes.cpp flplugin.cpp flshow.cpp                     flupdate.cpp frame.cpp frame.hpp global.cpp global.hpp                     grabber.cpp grpsort.cpp help.cpp hilight.cpp infolist.cpp                     interf.cpp keyboard.cpp language.cpp language.hpp                     lockscrn.cpp macro.cpp main.cpp manager.cpp manager.hpp                     message.cpp mix.cpp mkdir.cpp options.cpp panel.cpp                     plugapi.cpp plugins.cpp print.cpp qview.cpp rdrwdsk.cpp                     scrbuf.cpp scrbuf.hpp scrsaver.cpp setattr.cpp                     setcolor.cpp stddlg.cpp syslog.cpp treelist.cpp                     usermenu.cpp viewer.cpp vmenu.cpp Новые файлы       : Удаленные файлы   : Состав            : 00649.MoreNWZ.txt               ffolders.cpp.649.diff                     chgmmode.cpp.649.diff           cmdline.cpp.649.diff                     config.cpp.649.diff             copy.cpp.649.diff                     ctrlobj.cpp.649.diff            ctrlobj.hpp.649.diff                     delete.cpp.649.diff             dialog.cpp.649.diff                     edit.cpp.649.diff               editor.cpp.649.diff                     fileedit.cpp.649.diff           filelist.cpp.649.diff                     filepanels.cpp.649.diff         filetype.cpp.649.diff                     fileview.cpp.649.diff           filter.cpp.649.diff                     findfile.cpp.649.diff           flmodes.cpp.649.diff                     flplugin.cpp.649.diff           flshow.cpp.649.diff                     flupdate.cpp.649.diff           frame.cpp.649.diff                     frame.hpp.649.diff              global.cpp.649.diff                     global.hpp.649.diff             grabber.cpp.649.diff                     grpsort.cpp.649.diff            help.cpp.649.diff                     hilight.cpp.649.diff            infolist.cpp.649.diff                     interf.cpp.649.diff             keyboard.cpp.649.diff                     language.cpp.649.diff           language.hpp.649.diff                     lockscrn.cpp.649.diff           macro.cpp.649.diff                     main.cpp.649.diff               manager.cpp.649.diff                     manager.hpp.649.diff            message.cpp.649.diff                     mix.cpp.649.diff                mkdir.cpp.649.diff                     options.cpp.649.diff            panel.cpp.649.diff                     plugapi.cpp.649.diff            plugins.cpp.649.diff                     print.cpp.649.diff              qview.cpp.649.diff                     rdrwdsk.cpp.649.diff            scrbuf.cpp.649.diff                     scrbuf.hpp.649.diff             scrsaver.cpp.649.diff                     setattr.cpp.649.diff            setcolor.cpp.649.diff                     stddlg.cpp.649.diff             syslog.cpp.649.diff                     treelist.cpp.649.diff           usermenu.cpp.649.diff                     viewer.cpp.649.diff             vmenu.cpp.649.diff Основан на патче  : 648 Дополнение        :
1. Объединены ExecuteModal() и ExecuteModalPtr().

2. ReplaceCurrentFrame() заменена на более общую ReplaceFrame() (не
   тестировалось).

3. FrameManager оторван от CtrlObject.

4. Еще немножко борьбы с global.hpp (оттуда вынесены CtrlObject, Lang и
   ScrBuf).

5. Всякие чистки внутренностей Manager и CtrlObject.

6. Починены разные проблемы с перерисовкой при вызове редактора или
   вьюера из плагина (в частности, FARCmds).

+++ far/manager.cpp
@@ -5,10 +5,16 @@ manager.cpp
 
 */
 
-/* Revision: 1.15 11.05.2001 $ */
+/* Revision: 1.16 12.05.2001 $ */
 
 /*
 Modify:
+  12.05.2001 DJ
+    ! FrameManager оторван от CtrlObject
+    ! объединены ExecuteModal() и ExecuteModalPtr() (и о чем я думал, когда 
+      делал две функции?)
+    ! ReplaceCurrentFrame() заменена на более универсальную ReplaceFrame()
+      (с подачи ОТ)
   11.05.2001 OT
     ! Отрисовка Background
   10.05.2001 DJ
@@ -68,6 +74,9 @@ Modify:
 #include "panel.hpp"
 #include "savescr.hpp"
 #include "cmdline.hpp"
+#include "ctrlobj.hpp"
+
+Manager *FrameManager;
 
 Manager::Manager()
 {
@@ -236,14 +245,29 @@ void Manager::DeleteDestroyedFrame()
    заменить текущий фрейм на указанный (используется для обработки F6)
 */
 
-void Manager::ReplaceCurrentFrame (Frame *NewFrame)
+/* $ 12.05.2001 DJ
+   заменена на более общую функцию (с подачи ОТ)
+*/
+
+void Manager::ReplaceFrame (Frame *OldFrame, Frame *NewFrame)
 {
-  DestroyedFrame = FrameList [FramePos];
+  DestroyedFrame = OldFrame;
   DestroyedFrame->OnChangeFocus (0);
-  FrameToReplace = NewFrame;
+  if (OldFrame == CurrentFrame)
+    FrameToReplace = NewFrame;
+  else
+    for (int i=0; i<FrameCount; i++)
+      if (FrameList [i] == OldFrame)
+        FrameList [i] = NewFrame;
 }
 
 /* DJ $ */
+/* DJ $ */
+
+
+/* $ 10.05.2001 DJ
+   поддерживает ReplaceFrame (F6)
+*/
 
 int Manager::ExecuteModal (Frame &ModalFrame)
 {
@@ -251,36 +275,22 @@ int Manager::ExecuteModal (Frame &ModalFrame)
   AddFrame (&ModalFrame);
   ModalFrame.Show();
   DestroyedFrame = NULL;
-  while (DestroyedFrame != &ModalFrame)
-    ProcessMainLoop();
-  int exitCode = ModalFrame.GetExitCode();
-  DestroyedFrame = NULL;
-  return exitCode;
-}
-
-/* $ 10.05.2001 DJ
-   в отличие от ExecuteModal(), поддерживает ReplaceCurrentFrame (F6)
-*/
-
-int Manager::ExecuteModalPtr (Frame *ModalFrame)
-{
-  ModalSaveState();
-  AddFrame (ModalFrame);
-  ModalFrame->Show();
-  DestroyedFrame = NULL;
-  while (DestroyedFrame != ModalFrame)
+  Frame *CurModalFrame = &ModalFrame;
+  while (DestroyedFrame != CurModalFrame)
   {
     ProcessMainLoop();
     if (FrameToReplace)
     {
-      ModalFrame = FrameToReplace;
-      DeleteDestroyedFrame();
+      CurModalFrame = FrameToReplace;
+      if (DestroyedFrame == &ModalFrame)
+        DestroyedFrame = NULL;
+      else
+        DeleteDestroyedFrame();
       continue;
     }
   }
-  int exitCode = ModalFrame->GetExitCode();
-  DeleteDestroyedFrame();
-  return exitCode;
+  DestroyedFrame = NULL;
+  return CurModalFrame->GetExitCode();
 }
 
 /* DJ */
@@ -312,50 +322,6 @@ void Manager::NextFrame(int Increment)
   SetCurrentFrame (CurFrame);
 
   _D(SysLog("Manager::NextFrame(), set CurrentFrame=0x%p, %s",CurrentFrame,CurrentFrame->GetTypeName()));
-
-  if (CurFrame->GetType()==MODALTYPE_EDITOR)
-    UpdateRequired=TRUE;
-
-/*    int ExitCode=CurModal->GetExitCode();
-    if (ExitCode<=0)
-    {
-      delete CurModal;
-      for (int I=ModalPos+1;I<ModalCount;I++)
-        ModalList[I-1]=ModalList[I];
-      ModalCount--;
-      if (*NextName)
-      {
-        ActivateNextFrame();
-        return;
-      }
-    }
-    else
-    {
-      if (ExitCode==2)
-        SelectModal();
-      else
-        if (ExitCode==3)
-        {
-          if (--ModalPos<0)
-            ModalPos=ModalCount;
-        }
-        else
-          ModalPos++;
-      CurModal->Hide();
-    }
-*/
-//  } // while
-/*  CtrlObject->Cp()->ActivePanel->SetTitle();
-  if (UpdateRequired)
-  {
-    CtrlObject->Cp()->LeftPanel->Update(UPDATE_KEEP_SELECTION|UPDATE_SECONDARY);
-    CtrlObject->Cp()->RightPanel->Update(UPDATE_KEEP_SELECTION|UPDATE_SECONDARY);
-    UpdateRequired=TRUE;
-  }
-  CtrlObject->Cp()->LeftPanel->Redraw();
-  CtrlObject->Cp()->RightPanel->Redraw();
-  CtrlObject->CmdLine.Redraw();
-  CtrlObject->MainKeyBar.Redraw();*/
   _D(SysLog(-1));
 }
 
@@ -365,7 +331,7 @@ void Manager::SetCurrentFrame (Frame *NewCurFrame)
   CurrentFrame = NewCurFrame;
   CurrentFrame->ShowConsoleTitle();
   CurrentFrame->OnChangeFocus(1);
-  CtrlObject->Macro.SetMode(CurrentFrame->MacroMode);
+  CtrlObject->Macro.SetMode(CurrentFrame->GetMacroMode());
 }
